Bind details view model with details screen

+++ app/src/main/java/com/marvel/moviesapp/ui/composable/PaginatedMoviesList.kt
@@ -15,7 +15,7 @@ fun PaginatedMoviesList(
     modifier: Modifier = Modifier,
     favoriteState: (MovieModel) -> Unit,
     isFavoritesScreen: Boolean,
-    onItemClick: () -> Unit
+    onItemClick: (Int) -> Unit
 ) {
     LazyVerticalGrid(modifier = modifier, cells = GridCells.Fixed(3)) {
         items(list.itemCount) { index ->
